+++ CHANGES.md
@@ -1,5 +1,5 @@
 #### [unreleased]
-* convert to only use `iframe` and use Google Doc Viewer for Chrome
+* convert to only use `iframe` and use Google Doc Viewer for Chrome and mobile
 * dynamically render `iframe` in block based on browser
 
 #### 2.3.1 / 2023-09-10